+++ changelog.md
@@ -13,6 +13,7 @@
 - Implemented popup close button
 - Moved JS styles into SCSS file
 - Added an event listener that updates the MGRS coordinates when the mouse moves
+- Added pulsating prompt to remind the user to drag the symbol onto the map
 
 ---